vuln_backend/requirements.txt
Flask==2.2.5
Flask-Caching==1.11.1
Flask-Cors==3.0.10
Flask-Limiter==3.1.0
requests==2.31.0
validators==0.20.0
beautifulsoup4==4.12.2
builtwith==1.3.4
dnspython==2.4.2
pyOpenSSL==23.2.0
urllib3==2.0.7
flask[async]
